異常回復
エラーの解消
メモリリークの解消
タイトルの設置

+++ project/game/src/game/game_object/database/chr_db/chr_db.cpp
@@ -31,28 +31,28 @@ void cUnitDataBase::Load()
 	{
 		Temp.TexFilePath	= UnitDataCSV.GetString(i, 1);
 		Temp.Name			= UnitDataCSV.GetString(i, 2);
-		Temp.Life			= UnitDataCSV.GetInteger(i, 3);
-		Temp.Cooling		= UnitDataCSV.GetInteger(i, 4);
-		Temp.Battery		= UnitDataCSV.GetInteger(i, 5);
-		Temp.Parts		= UnitDataCSV.GetInteger(i, 6);
-		Temp.Ammo			= UnitDataCSV.GetInteger(i, 7);
-		Temp.Resist[0]	= UnitDataCSV.GetInteger(i, 8);
-		Temp.Resist[1]	= UnitDataCSV.GetInteger(i, 9);
-		Temp.Resist[2]	= UnitDataCSV.GetInteger(i, 10);
-		Temp.Inventory	= UnitDataCSV.GetInteger(i, 11);
-		Temp.HeadCount	= UnitDataCSV.GetInteger(i, 12);
-		Temp.ArmCount		= UnitDataCSV.GetInteger(i, 13);
-		Temp.ChestCount	= UnitDataCSV.GetInteger(i, 14);
-		Temp.BackCount	= UnitDataCSV.GetInteger(i, 15);
-		Temp.LegCount		= UnitDataCSV.GetInteger(i, 16);
-		Temp.ShlderCount	= UnitDataCSV.GetInteger(i, 17);
-		Temp.CardCount	= UnitDataCSV.GetInteger(i, 18);
+		Temp.Life			= (uint16_t)UnitDataCSV.GetInteger(i, 3);
+		Temp.Cooling		= (uint16_t)UnitDataCSV.GetInteger(i, 4);
+		Temp.Battery		= (uint16_t)UnitDataCSV.GetInteger(i, 5);
+		Temp.Parts		= (uint16_t)UnitDataCSV.GetInteger(i, 6);
+		Temp.Ammo			= (uint16_t)UnitDataCSV.GetInteger(i, 7);
+		Temp.Resist[0]	= (int16_t)UnitDataCSV.GetInteger(i, 8);
+		Temp.Resist[1]	= (int16_t)UnitDataCSV.GetInteger(i, 9);
+		Temp.Resist[2]	= (int16_t)UnitDataCSV.GetInteger(i, 10);
+		Temp.Inventory	= (uint8_t)UnitDataCSV.GetInteger(i, 11);
+		Temp.HeadCount	= (uint8_t)UnitDataCSV.GetInteger(i, 12);
+		Temp.ArmCount		= (uint8_t)UnitDataCSV.GetInteger(i, 13);
+		Temp.ChestCount	= (uint8_t)UnitDataCSV.GetInteger(i, 14);
+		Temp.BackCount	= (uint8_t)UnitDataCSV.GetInteger(i, 15);
+		Temp.LegCount		= (uint8_t)UnitDataCSV.GetInteger(i, 16);
+		Temp.ShlderCount	= (uint8_t)UnitDataCSV.GetInteger(i, 17);
+		Temp.CardCount	= (uint8_t)UnitDataCSV.GetInteger(i, 18);
 		for (int j = 0; j < 16; j++)
-			Temp.Equipped[j] = UnitDataCSV.GetInteger(i, j + 19);
+			Temp.Equipped[j] = (uint16_t)UnitDataCSV.GetInteger(i, j + 19);
 		for (int j = 0; j < 4; j++)
 		{
-			Temp.DropItemId[j]	= UnitDataCSV.GetInteger(i, j * 2 + 35);
-			Temp.DropRate[j]	= UnitDataCSV.GetInteger(i, j * 2 + 1 + 35);
+			Temp.DropItemId[j]	= (uint16_t)UnitDataCSV.GetInteger(i, j * 2 + 35);
+			Temp.DropRate[j]	= (uint8_t)UnitDataCSV.GetInteger(i, j * 2 + 1 + 35);
 		}
 		m_UnitDataList.push_back(Temp);
 	}
