+++ project/game/src/game/game_object/ui_manager/component/value/value.cpp
@@ -88,7 +88,7 @@ int cValue::GetNum(bool calcvalue)
 
 float cValue::GetFloatNum()
 {
-	return m_Value;
+	return (float)m_Value;
 }
 
 bool cValue::IsZero()
